requirements.txt
alibabacloud_dysmsapi20170525>=2.0.24
openpyxl>=3.1.2
python-dotenv>=1.0.0
click>=8.1.7

